requirements.txt
google-cloud-language
streamlit